navigation tests on login optios screen
